requirements.txt
python-dotenv==0.19.0
sqlalchemy==1.4.31
python-telegram-bot==13.7
